Книпки переделаны в Material. Все размеры и строки в ресурсах. Создан свой стиль для кнопок
