+++ app/src/main/res/values/dimens.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <dimen name="textSizeButton">30sp</dimen>
+    <dimen name="layout_width100dp">100dp</dimen>
+    <dimen name="layout_height70dp">70dp</dimen>
+    <dimen name="translationX_delete_one_element">-5dp</dimen>
+</resources>
