+++ app/src/main/res/values/themes.xml
@@ -13,4 +13,14 @@
         <item name="android:statusBarColor" tools:targetApi="l">?attr/colorPrimaryVariant</item>
         <!-- Customize your theme here. -->
     </style>
+
+    <style name="StyleButton" parent="Theme.Сalculator">
+        <item name="android:textSize">@dimen/textSizeButton</item>
+        <item name="android:background">@android:color/transparent</item>
+        <item name="android:textColor">@color/black</item>
+        <item name="android:layout_width">@dimen/layout_width100dp</item>
+        <item name="android:layout_height">@dimen/layout_height70dp</item>
+    </style>
+
+
 </resources>
